MA-8198 Streanming:video pause and progress bar moves to 0 when fullscreen
minUndequeuedBuffers affect setSurface if forced to 0. Keep it's
original value and modify cancelStart to satisfy vpu's requirement
of holding all buffers.

mEnqueuedBuffers is not the real buffer count in native window if forced
to 0 in flush state, because flushing won't dequeue all buffers from
native window.  It should be modified only when queuebuffer or
dequeuebuffer.

Change-Id: I6750a2f755e4693dc99fafb2949297c83ed48ede
Signed-off-by: mao.xu <mao.xu@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/1327
Reviewed-by: ming.zhou <ming.zhou@nxp.com>